@@ -1,105 +1,56 @@
 {
+  "name": "treat-monorepo",
   "private": true,
   "workspaces": [
-    "packages/*"
+    "packages/*",
+    "site",
+    "examples/*",
+    "test-helpers"
   ],
+  "preconstruct": {
+    "packages": [
+      "packages/treat",
+      "packages/react-treat"
+    ]
+  },
   "scripts": {
-    "pretest": "yarn build",
+    "lint": "tsc && manypkg check && prettier --check .",
+    "format": "prettier --write .",
     "test": "jest",
     "make-docs-manifest": "node ./site/makeDocsManifest.js",
     "prestart-docs": "yarn make-docs-manifest",
     "prebuild-docs": "yarn make-docs-manifest",
     "start-docs": "webpack-dev-server --config ./site/webpack.config.js",
     "build-docs": "NODE_ENV=production webpack --config ./site/webpack.config.js",
-    "build": "lerna run build",
-    "watch": "lerna run watch",
-    "release": "lerna publish --conventional-commits",
-    "commit": "git-cz",
+    "build": "preconstruct build",
+    "watch": "preconstruct watch",
+    "dev": "preconstruct dev",
+    "release": "echo 'TODO'",
     "debug-fixture": "node ./test-helpers/debugFixture.js"
   },
-  "config": {
-    "commitizen": {
-      "path": "cz-conventional-changelog"
-    }
-  },
-  "husky": {
-    "hooks": {
-      "commit-msg": "commitlint --edit --extends seek"
-    }
-  },
   "volta": {
-    "node": "10.15.2",
-    "yarn": "1.21.1"
-  },
-  "devDependencies": {
-    "@babel/cli": "^7.4.4",
-    "@babel/core": "^7.4.4",
-    "@babel/plugin-syntax-dynamic-import": "^7.2.0",
-    "@babel/preset-env": "^7.3.4",
-    "@babel/preset-react": "^7.0.0",
-    "@babel/preset-typescript": "^7.3.3",
-    "@commitlint/cli": "^7.6.1",
-    "@mdx-js/react": "^1.0.16",
-    "@types/classnames": "^2.2.7",
-    "@types/dedent": "^0.7.0",
-    "@types/hapi__joi": "^15.0.1",
-    "@types/jest": "^24.0.13",
-    "@types/jest-environment-puppeteer": "^4.0.0",
-    "@types/lodash": "^4.14.134",
-    "@types/lunr": "^2.3.2",
-    "@types/node": "^12.0.1",
-    "@types/puppeteer": "^1.12.4",
-    "@types/react": "^16.8.17",
-    "@types/react-dom": "^16.8.4",
-    "@types/react-highlight-words": "^0.16.0",
-    "@types/react-router-dom": "^4.3.3",
-    "@types/react-router-hash-link": "^1.2.1",
-    "babel-loader": "^8.0.5",
-    "basekick": "^3.1.1",
-    "bent": "^1.4.0",
-    "classnames": "^2.2.6",
-    "commitizen": "^3.1.1",
-    "commitlint-config-seek": "^1.0.0",
-    "concurrently": "^4.1.0",
-    "copy-webpack-plugin": "^5.0.3",
-    "cz-conventional-changelog": "^2.1.0",
-    "execa": "^3.4.0",
-    "express": "^4.16.4",
-    "fast-glob": "^2.2.7",
-    "file-loader": "^3.0.1",
-    "github-slugger": "^1.2.1",
-    "gray-matter": "^4.0.2",
-    "html-render-webpack-plugin": "^1.1.1",
-    "html-webpack-plugin": "^3.2.0",
-    "husky": "^2.2.0",
-    "jest": "^24.9.0",
-    "jest-puppeteer": "^4.1.0",
-    "lerna": "^3.13.4",
-    "lodash": "^4.17.11",
-    "lunr": "^2.3.6",
-    "mdx-loader": "^2.2.1",
-    "mime-types": "^2.1.22",
-    "mini-css-extract-plugin": "^0.8.0",
-    "next": "9.2.2",
-    "prettier": "^1.16.4",
-    "puppeteer": "^1.14.0",
-    "react": "^16.13.1",
-    "react-dom": "^16.13.1",
-    "react-head": "^3.1.1",
-    "react-highlight-words": "^0.16.0",
-    "react-router-dom": "^5.0.0",
-    "react-router-hash-link": "^1.2.1",
-    "remove-markdown": "^0.3.0",
-    "typescript": "^3.4.1",
-    "use-react-router": "^1.0.7",
-    "wait-for-localhost": "^3.2.0",
-    "webpack": "^4.30.0",
-    "webpack-cli": "^3.2.3",
-    "webpack-dev-server": "^3.4.1",
-    "webpack-merge": "^4.2.1",
-    "webpack-node-externals": "^1.7.2"
+    "node": "14.15.4",
+    "yarn": "1.22.10"
   },
   "resolutions": {
     "@types/react": "^16.8.17"
+  },
+  "dependencies": {
+    "@babel/core": "^7.12.10",
+    "@babel/preset-env": "^7.12.11",
+    "@babel/preset-react": "^7.12.10",
+    "@babel/preset-typescript": "^7.12.7",
+    "@manypkg/cli": "^0.17.0",
+    "@preconstruct/cli": "^2.0.1",
+    "@types/jest": "^26.0.20",
+    "@types/jest-environment-puppeteer": "^4.0.0",
+    "@types/puppeteer": "^1.12.4",
+    "babel-jest": "^26.6.3",
+    "babel-loader": "^8.2.2",
+    "jest": "^26.6.3",
+    "jest-puppeteer": "^4.1.0",
+    "prettier": "^2.2.1",
+    "puppeteer": "^1.14.0",
+    "typescript": "^4.1.3"
   }
 }
