@@ -1,14 +1,7 @@
 module.exports = {
   presets: [
     '@babel/preset-typescript',
-    [
-      '@babel/preset-env',
-      {
-        targets: {
-          node: 'current',
-        },
-      },
-    ],
     '@babel/preset-react',
+    ['@babel/preset-env', { targets: { node: '12' } }],
   ],
 };
