@@ -1,3 +1,3 @@
 module.exports = {
-    "presets": ["babel-preset-gatsby"]
-}
+  presets: ['babel-preset-gatsby'],
+};
