@@ -1,4 +1,6 @@
 import React from 'react';
-import {button} from './Button.treat'
+import { button } from './Button.treat';
 
-export const Button = ({...otherProps}) => <button {...otherProps} className={button}/>;
+export const Button = ({ ...otherProps }) => (
+  <button {...otherProps} className={button} />
+);
