@@ -1,16 +1,16 @@
-import {style} from 'treat';
-import {fontStack} from './fontStack';
+import { style } from 'treat';
+import { fontStack } from './fontStack';
 
 export const mesage = style({
-    fontFamily: fontStack
+  fontFamily: fontStack,
 });
 
 export const emoji = style({
-    display: 'inline-block',
-    fontSize: '3em',
-    selectors: {
-        '& + &': {
-            marginLeft: '1rem'
-        }
-    }
+  display: 'inline-block',
+  fontSize: '3em',
+  selectors: {
+    '& + &': {
+      marginLeft: '1rem',
+    },
+  },
 });
