@@ -1,11 +1,11 @@
-import {globalStyle} from 'treat';
+import { globalStyle } from 'treat';
 
 globalStyle('html', {
   margin: 0,
-  padding: 0
+  padding: 0,
 });
 
 globalStyle('body', {
   margin: '1em',
-  padding: 0
+  padding: 0,
 });
