@@ -6,17 +6,21 @@ export default () => {
   const [Component, setComponent] = useState(null);
 
   const handleShowMessage = () => {
-    import('../components/SecretMessage')
-      .then(
-        ({SecretMessage}) => setComponent(() => SecretMessage), 
-        error => console.error(error)
-      );
-  }
+    import('../components/SecretMessage').then(
+      ({ SecretMessage }) => setComponent(() => SecretMessage),
+      (error) => console.error(error),
+    );
+  };
 
   return (
     <>
-      <Button disabled={Boolean(Component)} onClick={handleShowMessage}>Show the secret message <span role="img" aria-label="shushing face">🤫</span></Button>
-      {Component && <Component/>}
+      <Button disabled={Boolean(Component)} onClick={handleShowMessage}>
+        Show the secret message{' '}
+        <span role="img" aria-label="shushing face">
+          🤫
+        </span>
+      </Button>
+      {Component && <Component />}
     </>
   );
-}
+};
