@@ -1,29 +1,34 @@
 import execa from 'execa';
 import * as path from 'path';
 import waitForLocalhost from 'wait-for-localhost';
-import getStyles from '../../../test-helpers/getStyles';
-import resolveBin from '../../../test-helpers/resolveBin';
+import getStyles from 'treat-test-helpers/getStyles';
+import resolveBin from 'treat-test-helpers/resolveBin';
 
 const gatsbyServerPort = 9999;
 const gatsbyBinaryPath = resolveBin('gatsby', 'gatsby');
-const gatsbyFixturePath = path.resolve(__dirname, '../../gatsby-plugin-treat-example');
+const gatsbyFixturePath = path.dirname(
+  require.resolve('gatsby-plugin-treat-example/package.json'),
+);
 const gatsbyExecArgs = {
-  shell: '/bin/bash', 
+  shell: '/bin/bash',
   cwd: gatsbyFixturePath,
   env: {
-    NODE_ENV: 'development'
+    NODE_ENV: 'development',
   },
-  extendEnv: true
+  extendEnv: true,
 };
 let gatsbyProcess;
 
 async function navigateToServerWhenReady() {
-  await waitForLocalhost({port: gatsbyServerPort});
+  await waitForLocalhost({ port: gatsbyServerPort });
   await page.goto(`http://localhost:${gatsbyServerPort}/`);
 }
 
 async function startDevServer() {
-  gatsbyProcess = execa(`${gatsbyBinaryPath} develop -p ${gatsbyServerPort}`, gatsbyExecArgs);
+  gatsbyProcess = execa(
+    `${gatsbyBinaryPath} develop -p ${gatsbyServerPort}`,
+    gatsbyExecArgs,
+  );
   gatsbyProcess.stdout.pipe(process.stdout);
   gatsbyProcess.stderr.pipe(process.stdout);
   await navigateToServerWhenReady();
@@ -31,7 +36,10 @@ async function startDevServer() {
 
 async function startProdServer() {
   await execa(`${gatsbyBinaryPath} build`, gatsbyExecArgs);
-  gatsbyProcess = execa(`${gatsbyBinaryPath} serve -p ${gatsbyServerPort}`, gatsbyExecArgs);
+  gatsbyProcess = execa(
+    `${gatsbyBinaryPath} serve -p ${gatsbyServerPort}`,
+    gatsbyExecArgs,
+  );
   gatsbyProcess.stdout.pipe(process.stdout);
   gatsbyProcess.stderr.pipe(process.stdout);
   await navigateToServerWhenReady();
@@ -43,9 +51,7 @@ async function stopServer() {
 }
 
 describe('gatsby', () => {
-
   describe('develop', () => {
-
     beforeAll(() => startDevServer());
     afterAll(() => stopServer());
 
@@ -53,13 +59,13 @@ describe('gatsby', () => {
       const styles = await getStyles(page, 'body');
       expect(styles).toMatchSnapshot();
     });
-    
+
     test('it renders some button styles', async () => {
       await page.waitForSelector('button');
       const styles = await getStyles(page, 'button');
       expect(styles).toMatchSnapshot();
     });
-    
+
     test('it loads and renders some span styles', async () => {
       await page.waitForSelector('button');
       await page.click('button');
@@ -67,11 +73,9 @@ describe('gatsby', () => {
       const styles = await getStyles(page, 'span');
       expect(styles).toMatchSnapshot();
     });
-
   });
-  
+
   describe('build', () => {
-    
     beforeAll(() => startProdServer());
     afterAll(() => stopServer());
 
@@ -79,13 +83,13 @@ describe('gatsby', () => {
       const styles = await getStyles(page, 'body');
       expect(styles).toMatchSnapshot();
     });
-    
+
     test('it renders some button styles', async () => {
       await page.waitForSelector('button');
       const styles = await getStyles(page, 'button');
       expect(styles).toMatchSnapshot();
     });
-    
+
     test('it loads and renders some span styles', async () => {
       await page.waitForSelector('button');
       await page.click('button');
@@ -93,7 +97,5 @@ describe('gatsby', () => {
       const styles = await getStyles(page, 'span');
       expect(styles).toMatchSnapshot();
     });
-
   });
-
 });
