@@ -38,11 +38,13 @@ export default ({ route, clientStats }: RenderParams) => {
   const assetPath = (filename: string) => `${publicPath}${filename}`;
   const assets = entrypoints.main.assets as Array<string>;
   const cssAssets = assets
-    .filter(asset => asset.endsWith('.css'))
-    .map(asset => `<link rel="stylesheet" href="${assetPath(asset)}"></link>`);
+    .filter((asset) => asset.endsWith('.css'))
+    .map(
+      (asset) => `<link rel="stylesheet" href="${assetPath(asset)}"></link>`,
+    );
   const jsAssets = assets
-    .filter(asset => asset.endsWith('.js'))
-    .map(asset => `<script src="${assetPath(asset)}"></script>`);
+    .filter((asset) => asset.endsWith('.js'))
+    .map((asset) => `<script src="${assetPath(asset)}"></script>`);
 
   const headTags: HeadTags = [];
   const html = render(route, headTags, basePath);
